refactor: move ResourceCardGroup
