+++ src/components/TeacherComponents/LessonShareCardGroup/LessonShareCardGroup.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent, FC } from "react";
 import { Control, Controller } from "react-hook-form";
 
-import ResourceCard from "@/components/DownloadAndShareComponents/ResourceCard";
+import ResourceCard from "@/components/TeacherComponents/ResourceCard";
 import { sortShareResources } from "@/components/DownloadAndShareComponents/helpers/sortResources";
 import { ResourceFormProps } from "@/components/DownloadAndShareComponents/downloadAndShare.types";
 import { LessonShareData, LessonShareSchema } from "@/node-lib/curriculum-api";
