+++ src/components/TeacherComponents/ResourceCard/ResourceCard.test.tsx
@@ -1,9 +1,10 @@
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import renderWithTheme from "../../../__tests__/__helpers__/renderWithTheme";
-
 import ResourceCard from "./ResourceCard";
+
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+
 
 describe("ResourceCard", () => {
   it("renders a checkbox with an icon, title and resource file type", () => {
